changed ID to use longs

+++ src/ecs/core/tiny_ecs.cpp
@@ -2,4 +2,4 @@
 #include "tiny_ecs.hpp"
 
 // All we need to store besides the containers is the id of every entity and callbacks to be able to remove entities across containers
-unsigned int Entity::id_count = 1;
+unsigned long Entity::id_count = 1;
